+++ src/Schema/AST/DocumentAST.php
@@ -17,7 +17,6 @@ use Nuwave\Lighthouse\Exceptions\DefinitionException;
 use Nuwave\Lighthouse\Exceptions\ParseException;
 use Nuwave\Lighthouse\Schema\Directives\ModelDirective;
 use Nuwave\Lighthouse\Support\Utils;
-use Serializable;
 
 /**
  * Represents the AST of the entire GraphQL schema document.
@@ -34,7 +33,7 @@ use Serializable;
  *     classNameToObjectTypeName: ClassNameToObjectTypeName,
  * }
  */
-class DocumentAST implements Serializable, Arrayable
+class DocumentAST implements Arrayable
 {
     const TYPES = 'types';
     const DIRECTIVES = 'directives';
@@ -221,27 +220,11 @@ class DocumentAST implements Serializable, Arrayable
     }
 
     /**
-     * @deprecated TODO remove in v6
-     */
-    public function serialize(): string
-    {
-        return serialize($this->__serialize());
-    }
-
-    /**
      * @param  SerializableArray  $data
      */
     public function __unserialize(array $data): void
     {
         $this->hydrateFromArray($data);
-    }
-
-    /**
-     * @deprecated TODO remove in v6
-     */
-    public function unserialize($data): void
-    {
-        $this->__unserialize(unserialize($data));
     }
 
     /**
@@ -259,8 +242,8 @@ class DocumentAST implements Serializable, Arrayable
         // Until they are accessed by name, they are kept in their array form.
 
         // @phpstan-ignore-next-line Since we start from the array form, the generic type does not match
-        $this->types = new NodeList($ast['types']);
+        $this->types = new NodeList($types);
         // @phpstan-ignore-next-line Since we start from the array form, the generic type does not match
-        $this->directives = new NodeList($ast['directives']);
+        $this->directives = new NodeList($directives);
     }
 }